+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-Development version
--------------------
+Version 1.1.1
+-------------
 
 * fixed pyws crashing if SOAP request body does not match the required schema,
 issue #16,